split datasource tests into 3 files
